pyproject.toml
[project]
name = "wuzup"
description = "A project for wuzup"
dynamic = ["version"]
authors = [{name = "csm10495", email = "csm10495@gmail.com"}]
license = {file = "LICENSE.md"}
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: POSIX",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4",
    "cairosvg",
    "pytesseract",
    "Pillow",
    "playwright",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[project.urls]
homepage = "https://github.com/csm10495/wuzup"
repository = "https://github.com/csm10495/wuzup"
documentation  = "https://csm10495.github.io/wuzup"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.scripts]
wuzup = "wuzup.cli:main"

[tool.setuptools]
packages = ["wuzup"]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "wuzup.__version__"}

[tool.pytest.ini_options]
pythonpath = [
    "."
]
markers = [
    "integration: tests requiring network access",
]

[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"
